+++ Twilio/Rest/Api/V2010/Account/ConferenceOptions.php
@@ -30,10 +30,12 @@ abstract class ConferenceOptions {
 
     /**
      * @param string $status The status
+     * @param string $announceUrl The announce_url
+     * @param string $announceMethod The announce_method
      * @return UpdateConferenceOptions Options builder
      */
-    public static function update($status = Values::NONE) {
-        return new UpdateConferenceOptions($status);
+    public static function update($status = Values::NONE, $announceUrl = Values::NONE, $announceMethod = Values::NONE) {
+        return new UpdateConferenceOptions($status, $announceUrl, $announceMethod);
     }
 }
 
@@ -166,9 +168,13 @@ class ReadConferenceOptions extends Options {
 class UpdateConferenceOptions extends Options {
     /**
      * @param string $status The status
+     * @param string $announceUrl The announce_url
+     * @param string $announceMethod The announce_method
      */
-    public function __construct($status = Values::NONE) {
+    public function __construct($status = Values::NONE, $announceUrl = Values::NONE, $announceMethod = Values::NONE) {
         $this->options['status'] = $status;
+        $this->options['announceUrl'] = $announceUrl;
+        $this->options['announceMethod'] = $announceMethod;
     }
 
     /**
@@ -179,6 +185,28 @@ class UpdateConferenceOptions extends Options {
      */
     public function setStatus($status) {
         $this->options['status'] = $status;
+        return $this;
+    }
+
+    /**
+     * The announce_url
+     * 
+     * @param string $announceUrl The announce_url
+     * @return $this Fluent Builder
+     */
+    public function setAnnounceUrl($announceUrl) {
+        $this->options['announceUrl'] = $announceUrl;
+        return $this;
+    }
+
+    /**
+     * The announce_method
+     * 
+     * @param string $announceMethod The announce_method
+     * @return $this Fluent Builder
+     */
+    public function setAnnounceMethod($announceMethod) {
+        $this->options['announceMethod'] = $announceMethod;
         return $this;
     }