+++ CHANGES.md
@@ -1,6 +1,15 @@
 twilio-php Changelog
 ====================
 
+[2018-02-09] Version 5.16.5
+----------------------------
+**Api**
+- Add `AnnounceUrl` and `AnnounceMethod` params for conference announce
+
+**Chat**
+- Add support to looking up user channels by identity in v1
+
+
 [2018-01-30] Version 5.16.4
 ----------------------------
 **Api**